Make brightness an int.

+++ .config/powerline/segments/common/brightness.py
@@ -5,5 +5,6 @@ from powerline.lib.shell import run_cmd
 
 def brightness(pl):
     '''Return the monitor's brightness according to xbacklight.'''
-    brightness = round(float(run_cmd(pl, ["xbacklight", "-get"])), 0)
+    # It doesn't want to go right from float to int.
+    brightness = int(round(float(run_cmd(pl, ["xbacklight", "-get"])), 0))
     return "{}%".format(str(brightness))
